[BOOK-434] refactor: DataGraph AppGraph 책임 분리
각 :core 모듈별 그래프 생성하는 방식으로 변경
