+++ app/src/main/kotlin/com/ninecraft/booket/di/AppGraph.kt
@@ -5,16 +5,6 @@ import android.app.Application
 import android.app.Service
 import android.content.Context
 import com.ninecraft.booket.core.di.DataScope
-import com.google.firebase.Firebase
-import com.google.firebase.analytics.FirebaseAnalytics
-import com.google.firebase.analytics.analytics
-import com.google.firebase.installations.FirebaseInstallations
-import com.google.firebase.installations.installations
-import com.google.firebase.messaging.FirebaseMessaging
-import com.google.firebase.messaging.messaging
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import com.google.firebase.remoteconfig.remoteConfig
-import com.google.firebase.remoteconfig.remoteConfigSettings
 import com.slack.circuit.foundation.Circuit
 import com.slack.circuit.runtime.presenter.Presenter
 import com.slack.circuit.runtime.ui.Ui
@@ -55,27 +45,6 @@ interface AppGraph {
             .setAnimatedNavDecoratorFactory(CrossFadeNavDecoratorFactory())
             .build()
     }
-
-    @Provides
-    fun provideFirebaseRemoteConfig(): FirebaseRemoteConfig {
-        return Firebase.remoteConfig.apply {
-            val configSettings by lazy {
-                remoteConfigSettings {
-                    minimumFetchIntervalInSeconds = if (com.ninecraft.booket.BuildConfig.DEBUG) 0 else 60
-                }
-            }
-            setConfigSettingsAsync(configSettings)
-        }
-    }
-
-    @Provides
-    fun provideFirebaseMessaging(): FirebaseMessaging = Firebase.messaging
-
-    @Provides
-    fun provideFirebaseInstallations(): FirebaseInstallations = Firebase.installations
-
-    @Provides
-    fun provideFirebaseAnalytics(): FirebaseAnalytics = Firebase.analytics
 
     val circuit: Circuit
