+++ core/data/impl/src/main/kotlin/com/ninecraft/booket/core/data/impl/repository/DefaultUserRepository.kt
@@ -21,7 +21,7 @@ import kotlinx.coroutines.tasks.await
 
 @SingleIn(DataScope::class)
 @Inject
-internal class DefaultUserRepository(
+class DefaultUserRepository(
     private val service: ReedService,
     private val onboardingDataSource: OnboardingDataSource,
     private val notificationDataSource: NotificationDataSource,
