+++ core/datastore/impl/src/main/kotlin/com/ninecraft/booket/core/datastore/impl/di/DataStoreGraph.kt
@@ -1,0 +1,90 @@
+package com.ninecraft.booket.core.datastore.impl.di
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+import com.ninecraft.booket.core.datastore.api.datasource.BookRecentSearchDataSource
+import com.ninecraft.booket.core.datastore.api.datasource.LibraryRecentSearchDataSource
+import com.ninecraft.booket.core.datastore.api.datasource.NotificationDataSource
+import com.ninecraft.booket.core.datastore.api.datasource.OnboardingDataSource
+import com.ninecraft.booket.core.datastore.api.datasource.TokenDataSource
+import com.ninecraft.booket.core.datastore.impl.datasource.DefaultBookRecentSearchDataSource
+import com.ninecraft.booket.core.datastore.impl.datasource.DefaultLibraryRecentSearchDataSource
+import com.ninecraft.booket.core.datastore.impl.datasource.DefaultNotificationDataSource
+import com.ninecraft.booket.core.datastore.impl.datasource.DefaultOnboardingDataSource
+import com.ninecraft.booket.core.datastore.impl.datasource.DefaultTokenDataSource
+import com.ninecraft.booket.core.di.DataScope
+import dev.zacsweers.metro.ContributesTo
+import dev.zacsweers.metro.Provides
+
+private const val TOKEN_DATASTORE_NAME = "TOKENS_DATASTORE"
+private const val BOOK_RECENT_SEARCH_DATASTORE_NAME = "BOOK_RECENT_SEARCH_DATASTORE"
+private const val LIBRARY_RECENT_SEARCH_DATASTORE_NAME = "LIBRARY_RECENT_SEARCH_DATASTORE"
+private const val ONBOARDING_DATASTORE_NAME = "ONBOARDING_DATASTORE"
+private const val NOTIFICATION_DATASTORE_NAME = "NOTIFICATION_DATASTORE"
+
+private val Context.tokenDataStore by preferencesDataStore(name = TOKEN_DATASTORE_NAME)
+private val Context.bookRecentSearchDataStore by preferencesDataStore(name = BOOK_RECENT_SEARCH_DATASTORE_NAME)
+private val Context.libraryRecentSearchDataStore by preferencesDataStore(name = LIBRARY_RECENT_SEARCH_DATASTORE_NAME)
+private val Context.onboardingDataStore by preferencesDataStore(name = ONBOARDING_DATASTORE_NAME)
+private val Context.notificationDataStore by preferencesDataStore(name = NOTIFICATION_DATASTORE_NAME)
+
+@ContributesTo(DataScope::class)
+interface DataStoreGraph {
+
+    @TokenDataStore
+    @Provides
+    fun provideTokenDataStore(
+        context: Context,
+    ): DataStore<Preferences> = context.tokenDataStore
+
+    @BookRecentSearchDataStore
+    @Provides
+    fun provideBookRecentSearchDataStore(
+        context: Context,
+    ): DataStore<Preferences> = context.bookRecentSearchDataStore
+
+    @LibraryRecentSearchDataStore
+    @Provides
+    fun provideLibraryRecentSearchDataStore(
+        context: Context,
+    ): DataStore<Preferences> = context.libraryRecentSearchDataStore
+
+    @OnboardingDataStore
+    @Provides
+    fun provideOnboardingDataStore(
+        context: Context,
+    ): DataStore<Preferences> = context.onboardingDataStore
+
+    @NotificationDataStore
+    @Provides
+    fun provideNotificationDataStore(
+        context: Context,
+    ): DataStore<Preferences> = context.notificationDataStore
+
+    @Provides
+    fun provideTokenDataSource(
+        defaultTokenDataSource: DefaultTokenDataSource,
+    ): TokenDataSource = defaultTokenDataSource
+
+    @Provides
+    fun provideBookRecentSearchDataSource(
+        defaultBookRecentSearchDataSource: DefaultBookRecentSearchDataSource,
+    ): BookRecentSearchDataSource = defaultBookRecentSearchDataSource
+
+    @Provides
+    fun provideLibraryRecentSearchDataSource(
+        defaultLibraryRecentSearchDataSource: DefaultLibraryRecentSearchDataSource,
+    ): LibraryRecentSearchDataSource = defaultLibraryRecentSearchDataSource
+
+    @Provides
+    fun provideOnboardingDataSource(
+        defaultOnboardingDataSource: DefaultOnboardingDataSource,
+    ): OnboardingDataSource = defaultOnboardingDataSource
+
+    @Provides
+    fun provideNotificationDataSource(
+        defaultNotificationDataSource: DefaultNotificationDataSource,
+    ): NotificationDataSource = defaultNotificationDataSource
+}
